Fix: Cambio en la persistencia de la API

+++ src/main/java/com/servicio2/Reconocimientos/Extra/DTOs/DTOColaboradorSolicitado.java
@@ -12,5 +12,5 @@ public class DTOColaboradorSolicitado {
         private String apellido;
         private String email;
         private Double puntaje;
-        private Long cantidadDonaciones;
+        private Integer cantidadDonaciones;
 }
